쿠폰 이벤트 및 사용자 관련 로직 리팩토링: 쿠폰 이벤트 생성 및 발급 로직 개선, DTO 및 예외 처리 수정, 테스트 케이스 업데이트. 재고 관리 및 유효성 검사를 위한 메서드 추가.

+++ src/main/kotlin/kr/hhplus/be/server/controller/couponevent/CouponEventController.kt
@@ -21,7 +21,7 @@ class CouponEventController(
         val criteria = CreateCouponEventCriteria(
             benefitMethod = req.benefitMethod,
             benefitAmount = req.benefitAmount,
-            totalIssueAmount = req.totalIssueAmount
+            totalIssueAmount = req.totalIssueAmount.toLong()
         )
         
         val result = couponEventFacade.createCouponEvent(criteria)
